Introduce common library functions
Reading file data and parsing from markdown proved to be an incredibly
repetitive task. I've refactored the functionality into it's own module
called "common" under "lib", and in doing so I've drastically cleaned
up the code in each individual library module.

I had to update some pages to run their library functions
asynchronously, and I've also found that a few pages were loading
projects and passing that information to the navbar, which was for an
outdated design and is no longer necessary.